enable oxlint import/no-cycle
lint for import cycles.

- moved `subscribe()` from `cache.ts -> subscribe.ts` 
- moved `writeData` from `cache.ts -> writeData.ts`
- moved `getOrCreateCacheForArtifact` from `cache.ts ->
  getOrCreateCacheForArtifact.ts`
- moved `stableCopy` and `isArray` from `cache.ts -> util.ts`
- moved `maybeUnwrapNetworkRequest` from `useResult ->
  maybeUnwrapNetworkRequest.ts`
- removed a nasty cycle `read -> componentCache -> useReadAndSubscribe
  -> read`, it required splitting `componentCache`, this is dependency of core on react  `core -> react -> core` used #564
