+++ libs/isograph-react/vitest.config.ts
@@ -26,4 +26,9 @@ export default defineProject({
       },
     }),
   ],
+  resolve: {
+    alias: {
+      '@iso': './tests/__isograph/iso.ts',
+    },
+  },
 });
